Starbucks Lid v3
There are a couple of changes with this version. First, I finally have the lid
in a shape that I think evokes the feel of a Starbucks lid. Second, I realized
that I converted Strokes to Paths in Inkscape and what I should have done was
converted Objects to Paths. The original conversion created too many paths and
the new one is smaller and cuts faster.